New makefile that allows for building both versions
Both random-walk (with OpenMP support) and graphics can now be built by
running make all. Although that still requires that both Clang (with
alias g++) and GCC (with alias g++-4.9) are installed. Each of
random-walk and graphics can be built separately by calling make
random-walk or make graphics. Those will only require the compiler for
that target to be installed.