Match nodes assigned to cores with METIS partition
This introduces an (optional) first line of TGF files which describes
the partitioning of the graph. If this line exists, it must start with
'METIS', which is follow n+1 numbers to describe the limits for each
partition. The first number is always 0. The rest are ordered lower
(inclusive) and upper (exclusive) limits.

This function will only be activated if the random-walk program is
called with a -p parameter taht matches that partitioning that is
stored in the file.